refactor(dto): change `speed` field type to `BigDecimal` in `WeatherApiResponseDto`

+++ src/main/java/org/as1iva/dto/response/WeatherApiResponseDto.java
@@ -68,7 +68,7 @@ public class WeatherApiResponseDto {
     @JsonIgnoreProperties(ignoreUnknown = true)
     public static class WindDto {
 
-        private Double speed;
+        private BigDecimal speed;
     }
 
     @Getter
